expand dataset volume

+++ copy_dataset.py
@@ -2,7 +2,9 @@
 import os
 import shutil
 
-dataset_path_list = ["E:\\PythonCodes\\bbox3d_annotation_tools\\session0_center_data"]
+dataset_path_list = ["E:\\PythonCodes\\bbox3d_annotation_tools\\session0_center_data", 
+                     "E:\\PythonCodes\\bbox3d_annotation_tools\\session0_right_data", 
+                     "D:\\3Dmark2\\session6_right_data"]
 voc_img_dir = "E:\\PythonCodes\\bbox3d_recognition_v2\\DATAdevkit\\DATA2021\\JPEGImages"
 voc_anno_dir = "E:\\PythonCodes\\bbox3d_recognition_v2\\DATAdevkit\\DATA2021\\Annotations"
 
@@ -18,11 +20,13 @@ for single_dataset_path_list in dataset_path_list:
             xml_raw_path = single_dataset_path_list + "\\" + file_name
             xml_new_path = voc_anno_dir + "\\" + file_name
             # copy xml
-            shutil.copy(xml_raw_path, xml_new_path)
+            if not os.path.exists(xml_new_path):
+                shutil.copy(xml_raw_path, xml_new_path)
 
             img_raw_path = single_dataset_path_list + "\\" + file_name[:-4] + ".jpg"
             img_new_path = voc_img_dir + "\\" + file_name[:-4] + ".jpg"
             # copy img
-            shutil.copy(img_raw_path, img_new_path)
+            if not os.path.exists(img_new_path):
+                shutil.copy(img_raw_path, img_new_path)
 
 print("finish copy imgs and annotations!")
